+++ capi/build.rs
@@ -129,6 +129,12 @@ fn cc(src: &Path, dst: &str, options: &[&str]) {
 }
 
 fn main() {
+    let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap();
+    if target_os == "linux" || target_os == "android" {
+        println!("cargo:rustc-cfg=linux");
+    }
+    println!("cargo:rustc-check-cfg=cfg(linux)");
+
     let crate_dir = PathBuf::from(env::var_os("CARGO_MANIFEST_DIR").unwrap());
 
     #[cfg(feature = "generate-c-header")]